On branch master  Your branch is ahead of 'origin/master' by 1 commit.  Changes to be committed: 	modified:   Client_Socket/Client_Socket.hpp 	modified:   Server_Socket/Server_Socket.hpp

+++ Client_Socket/Client_Socket.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstddef>
 #ifndef CLIENT_SOCKET
 #define CLIENT_SOCKET
 #endif
@@ -50,11 +51,11 @@ namespace Client_Socket_NSP
         }
         // 接收指定长度内容先保存至缓冲区(注意缓冲区大小)，然后拷贝一份到rbuffer并清空缓冲区
         // 默认接收_buffersize长度
-        int Recv(uint32_t read_len)
+        size_t Recv(int read_len)
         {
             if (read_len <= 0)
                 read_len = this->_buffersize;
-            int len = recv(this->client_fd, this->_buffer, read_len, 0);
+            size_t len = recv(this->client_fd, this->_buffer, read_len, 0);
             this->rbuffer += this->_buffer;
             memset(this->_buffer, 0, this->_buffersize);
             return len;
@@ -67,11 +68,11 @@ namespace Client_Socket_NSP
 
         // 发送wbuffer中的内容并清空wbuffer已发送的内容
         // 默认发送wbuffer中的所有内容
-        int Send(uint32_t send_len)
+        size_t Send(int send_len)
         {
             if (send_len <= 0)
                 send_len = this->wbuffer.length();
-            int len = send(this->client_fd, wbuffer.c_str(), send_len, 0);
+            size_t len = send(this->client_fd, wbuffer.c_str(), send_len, 0);
             Erase(len);
             return len;
         }
